+++ src/Pim/Bundle/FilterBundle/Filter/ProductCompletenessFilter.php
@@ -2,9 +2,12 @@
 
 namespace Pim\Bundle\FilterBundle\Filter;
 
+use Symfony\Component\Form\FormFactoryInterface;
+use Oro\Bundle\FilterBundle\Filter\FilterUtility;
 use Oro\Bundle\FilterBundle\Filter\BooleanFilter;
 use Oro\Bundle\FilterBundle\Form\Type\Filter\BooleanFilterType;
 use Oro\Bundle\FilterBundle\Datasource\FilterDatasourceAdapterInterface;
+use Pim\Bundle\CatalogBundle\Entity\Repository\ProductRepository;
 
 /**
  * Overriding of boolean filter to filter by the product completeness
@@ -13,8 +16,30 @@ use Oro\Bundle\FilterBundle\Datasource\FilterDatasourceAdapterInterface;
  * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class CompletenessFilter extends BooleanFilter
+class ProductCompletenessFilter extends BooleanFilter
 {
+    /**
+     * @var ProductRepository
+     */
+    protected $repository;
+
+    /**
+     * Constructor
+     *
+     * @param FormFactoryInterface $factory
+     * @param FilterUtility        $util
+     * @param ProductRepository    $repository
+     */
+    public function __construct(
+        FormFactoryInterface $factory,
+        FilterUtility $util,
+        ProductRepository $repository
+    ) {
+        parent::__construct($factory, $util);
+
+        $this->repository = $repository;
+    }
+
     /**
      * {@inheritdoc}
      */
@@ -25,8 +50,11 @@ class CompletenessFilter extends BooleanFilter
             return false;
         }
 
-        $qb    = $ds->getQueryBuilder();
-        $field = 'completeness.ratio';
+        $qb        = $ds->getQueryBuilder();
+        $rootAlias = $qb->getRootAlias();
+        $joinAlias = 'filterCompleteness';
+        $field     = $joinAlias.'.ratio';
+        $this->repository->addCompleteness($qb, $joinAlias);
 
         switch ($data['value']) {
             case BooleanFilterType::TYPE_YES:
@@ -34,10 +62,7 @@ class CompletenessFilter extends BooleanFilter
                 break;
             case BooleanFilterType::TYPE_NO:
             default:
-                $expression = $qb->expr()->orX(
-                    $qb->expr()->neq($field, '100'),
-                    $qb->expr()->isNull($field)
-                );
+                $expression = $qb->expr()->lt($field, '100');
                 break;
         }