+++ src/Pim/Bundle/FilterBundle/Resources/public/js/datafilter/filter/product_completeness-filter.js
@@ -10,8 +10,8 @@ define(
          * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
          * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
          *
-         * @export  oro/datafilter/completeness-filter
-         * @class   oro.datafilter.CompletenessFilter
+         * @export  oro/datafilter/product_completeness-filter
+         * @class   oro.datafilter.ProductCompletenessFilter
          * @extends oro.datafilter.SelectFilter
          */
         return SelectFilter.extend({});